Move AppDriver to androidTestShared folder.
